+++ apps/web/components/Charts/MixedChart/MixedChart.tsx
@@ -17,7 +17,7 @@ import {
   CustomTooltip,
   CustomizedRightAxisTick,
   YAxisLabel,
-} from '../utils'
+} from '../sharedChartComponents'
 import { Box } from '@island.is/island-ui/core'
 
 interface GraphDataProps {
